Module pallet_assets_chain_extension::traits

Contains all stuff related to the PalletAssets trait. You can be imported into the ink and substrate codebase.

Enums

Error

The common errors that can be emitted by the pallet-asset.

Origin

The origin of the call. The smart contract can execute methods on behalf of the caller or itself.

Traits

Environment

The trait describes types used in the chain extension definition. Substrate and ink! side can have its types, so the trait is agnostic.

PalletAssets