Trait pallet_assets_chain_extension::traits::Environment

pub trait Environment {
    type AccountId;
    type AssetId;
    type Balance;
}

The trait describes types used in the chain extension definition. Substrate and ink! side can have its types, so the trait is agnostic.

Required Associated Types

type AccountId

type AssetId

type Balance